requirements.txt
google-cloud-aiplatform>=1.49.0
pandas
scikit-learn
joblib
gcsfs
